Add files via upload
The policy I’m addressing is the number of Drug Activity complaints in NYC (year 2020), more specifically by incident zip codes. I created a graduated map of the complaints by incident zip codes then added a NYC District Public Health Office (DPHO) Neighborhoods shapefile to see the location of DPHOs. My analysis shows that DPHOs are located in areas that experience the highest number of drug activity complaints. However, this poses some serious lingering questions about the effectiveness of DPHOs in their respective regions.

From one perspective, the DPHOs being located in the highest areas of drug activity could mean that DPHOs are not doing a well enough job in addressing drug activity within the surrounding areas. If this is the case, then DPHOs need to reassess their objectives and policy implementations/executions. Furthermore, I strongly believe that there is a severe lack of DPHOs in NYC, which could translate to as why they happen to be in areas of high drug activity because there is not enough of them to address drug use within their regions.